+++ modules/040-control-plane-manager/hooks/resources/hook_autotune_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package hooks
+package resources
 
 import (
 	"context"
